+++ packages/core/src/lib/bus.ts
@@ -193,7 +193,9 @@ export class Bus {
   }
 
   private async loadRelatedApps (app: App) {
-    await Promise.all(app.relatedApps.map(({ name, ctx }) => this.loadApp(name, ctx)))
+    for (const { name, ctx } of app.relatedApps) {
+      await this.loadApp(name, ctx)
+    }
   }
 
   /**