+++ packages/core/src/lib/loader.ts
@@ -2,16 +2,21 @@ import { ScriptType, LinkType } from '../types'; // eslint-disable-line
 
 export const loadScript = async (scriptDeclare: ScriptType) => {
   const promise: Promise<void> = new Promise(resolve => {
-    const scriptAttrs: ScriptType = typeof scriptDeclare !== 'string'
-      ? scriptDeclare
-      : {
-        type: 'text/javascript',
-        src: scriptDeclare
-      }
-    const script = document.createElement('script')
-    Object.entries(scriptAttrs).forEach(([attr, value]) => {
-      script[attr] = value
-    })
+    let script: HTMLScriptElement = null
+    if (scriptDeclare instanceof HTMLScriptElement) {
+      script = scriptDeclare
+    } else {
+      script = document.createElement('script')
+      const scriptAttrs: ScriptType = typeof scriptDeclare !== 'string'
+        ? scriptDeclare
+        : {
+          type: 'text/javascript',
+          src: scriptDeclare
+        }
+      Object.entries(scriptAttrs).forEach(([attr, value]) => {
+        script[attr] = value
+      })
+    }
     if (script.src) {
       script.onload = script.onerror = () => {
         resolve()
@@ -26,17 +31,22 @@ export const loadScript = async (scriptDeclare: ScriptType) => {
 }
 
 export const loadLink = (linkDeclare: LinkType) => {
-  const linkAttrs: LinkType = typeof linkDeclare !== 'string'
-    ? linkDeclare
-    : {
-      rel: 'stylesheet',
-      type: 'text/css',
-      href: linkDeclare
-    }
-  const link = document.createElement('link')
-  Object.entries(linkAttrs).forEach(([attr, value]) => {
-    link[attr] = value
-  })
+  let link: HTMLLinkElement = null
+  if (linkDeclare instanceof HTMLLinkElement) {
+    link = linkDeclare
+  } else {
+    const linkAttrs: LinkType = typeof linkDeclare !== 'string'
+      ? linkDeclare
+      : {
+        rel: 'stylesheet',
+        type: 'text/css',
+        href: linkDeclare
+      }
+    link = document.createElement('link')
+    Object.entries(linkAttrs).forEach(([attr, value]) => {
+      link[attr] = value
+    })
+  }
   document.head.appendChild(link)
 }