+++ packages/playground/apps/react-app/lifecycles.tsx
@@ -4,9 +4,9 @@ import App from './components/App'
 let rootContainer: HTMLElement = null
 
 export const onBootstrap = (container: HTMLElement) => {
-  rootContainer = container
+  rootContainer = container ?? document.getElementById('react-app')
   // @ts-ignore
-  ReactDOM.render(<App />, container) // TODO: remove @ts-ignore
+  ReactDOM.render(<App />, rootContainer) // TODO: remove @ts-ignore
 }
 
 export const onDestroy = () => {